Add new logincontroller (3 new API), Configured spring security to custom login page successfully, added template static folders (to follow best standards for web application developement)
