+++ src/main/resources/templates/login.html
@@ -1,24 +1,26 @@
 <!DOCTYPE html>
-<html>
+<html xmlns = "http://www.w3.org/1999/xhtml" 
+xmlns:th = "http://www.thymeleaf.org">
     <head>
         <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
         <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-        <link href="./login.css" rel="stylesheet" type="text/css">
+        <link href="../css/login.css" rel="stylesheet" type="text/css">
         <title>Spotify Login</title>
     </head>
     <body>
         <header>Login</header>
-        <form action="/login" method="POST">
+        <p th:if=${loginError} class="error">Wrong login credentials</p>
+        <form th:action="@{/login}" method="POST">
             <div id="email-container">
                 <label for="username">Email: </label>
-                <input id="username" type="email" name="username" required>
+                <input id="username" name="username" autofocus required>
             </div>
             <div id="password-container">
                 <label for="password">Password: </label>
                 <input id="password" type="password" name="password" required>
             </div>
-            <input type="submit" id="submit" value="Login"/>
-            <a href="http://localhost:8080/register.html" class="btn-link">Register</a>
+            <input type="submit" id="submit" value="Log in"/>
+            <a href="/register" class="btn-link">Register</a>
             <div id="forgot-password">
                 <a href="#" class="btn-non-link">Forgot your password?</a>
             </div>
