+++ src/main/webapp/app.module.js
@@ -30,8 +30,8 @@ angular
             })
             .when('/playlist',
             {
-                // controller: 'PlaylistController',
-                // controllerAs: 'pCtrl',
+                controller: 'PlaylistController',
+                controllerAs: 'pCtrl',
                 templateUrl: 'partials/playlist.html'
             })
             .otherwise({
