+++ src/main/webapp/partials/playlist.html
@@ -1,4 +1,4 @@
-<div class="ng-container" ng-controller="PlaylistController as pCtrl">
+<div class="ng-container">
     <header>
         <h1 id="page-title">Playlists</h1>
     </header>
