+++ src/main/webapp/services/myCache.service.js
@@ -1,0 +1,10 @@
+(function() {
+    angular
+    .module('spotifyWebApp')
+    .factory('MyCacheService', ['$cacheFactory', MyCacheService]);
+
+    function MyCacheService($cacheFactory) {
+        return $cacheFactory('myData');   
+    }
+    
+})();
